requirements/base.txt
# Base requirements

# Dash and friends.
dash[testing]
dash-renderer 
dash-html-components 
dash-core-components 
dash-bootstrap-components
dash_table 
plotly

# Numpy, Pandas and friends.
numpy
pandas
matplotlib

# PySAM
pysam ==0.15.3

# Interval arithmetics
intervaltree ==3.0.2

# Natural sorting.
natsort ==6.0.0

# Caching functionality for Flask.
flask-caching

# Lightweight setup of data classes.
attrs
# Simplified logging.
logzero
# Support for YAML access.
ruamel.yaml
# Unified access to files within archives or behind protocols.
fs
fs.sshfs >=0.11.0
s3fs  # NOT fs-s3fs
# HTTP requests
requests
# Parsing of HTML listings
htmllistparse
# TODO: we need an fs[._]irods
# Access to files through iRODS.
python-irodsclient
